+++ src/hooks/orderConfirmation/useCoupons/useCoupons.ts
@@ -1,5 +1,5 @@
-import fetchCoupons from "../../apis/fetchCoupons";
-import useFetchData from "../useFetchData";
+import fetchCoupons from "../../../apis/fetchCoupons";
+import useFetchData from "../../useFetchData";
 
 const useCoupons = () => {
   const { data: coupons } = useFetchData({
